# backend/swagger-spec.json
{"openapi":"3.0.0","paths":{"/posts":{"post":{"operationId":"PostsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostDTO"}}}}},"tags":["post"]},"get":{"operationId":"PostsController_findAll","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostDTO"}}}}}},"tags":["post"]}},"/posts/{id}":{"get":{"operationId":"PostsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostDTO"}}}}},"tags":["post"]},"patch":{"operationId":"PostsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePostDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostDTO"}}}}},"tags":["post"]},"delete":{"operationId":"PostsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["post"]}},"/database-files/{id}":{"get":{"operationId":"DatabaseFilesController_getDatabaseFileById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"content":{"image":{"schema":{"type":"string","format":"binary"}}},"description":""}},"tags":["database-file"]}},"/database-files":{"post":{"operationId":"DatabaseFilesController_uploadDatabaseFile","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DatabaseFileDTO"}}}}},"tags":["database-file"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDTO"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}}},"info":{"title":"Posts API","description":"","version":"1.0.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:8000"}],"components":{"schemas":{"CreatePostDto":{"type":"object","properties":{"title":{"type":"string"},"text":{"type":"string"},"previewId":{"type":"number"}},"required":["title","text"]},"PostDTO":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"text":{"type":"string"},"previewId":{"type":"number"}},"required":["id","title","text","previewId"]},"UpdatePostDto":{"type":"object","properties":{"title":{"type":"string"},"text":{"type":"string"},"previewId":{"type":"number"}}},"Uint8Array":{"type":"object","properties":{}},"DatabaseFileDTO":{"type":"object","properties":{"id":{"type":"number"},"filename":{"type":"string"},"data":{"$ref":"#/components/schemas/Uint8Array"}},"required":["id","filename","data"]},"AuthDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]}}}}